requirements.txt
loguru==0.6.0
openai==0.27.2
python-dotenv==1.0.0
setuptools==63.2.0
